fix: no providers usage in root.
